feat: add effect and curse selection menus for Reliquary
- Implemented effect selection menu with dynamic category filtering and search functionality.
- Created curse selection menu with similar features, allowing users to pick curses based on categories.
- Introduced utility functions for gradient theming and category theme mapping to enhance UI consistency.
- Removed empty Test.txt file from downloads directory.
